refactor(junction-tree): refactor junction-tree structure
